chore(readme): add deprecation notice for Angular V17
The package is designed for Material Legacy (pre-version 17) and will not be updated to support Angular V17. The decision to discontinue support stems from disagreements with Angular V17's new syntax and the overall direction in which Angular is heading. Users are advised to refer to the README for a detailed explanation and recommendations.